--- app/controllers/AboutController.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace app\controllers;
-
-class AboutController extends Controller
-{
-    public function index()
-    {
-        $title = APP_NAME . ' - ' . 'About';
-        $view = "AboutView";
-        require_once VIEWS_PATH . "MasterView.php";
-    }
-}
