--- app/controllers/SignInController.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace app\controllers;
-
-class SignInController extends Controller
-{
-    public function index()
-    {
-    }
-}
