--- app/controllers/SignUpController.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace app\controllers;
-
-class SignUpController extends Controller
-{
-    public function index()
-    {
-    }
-}
