--- app/views/AboutView.php
@@ -1,5 +1,0 @@
-<div class="container text-center d-flex w-100 h-100 justify-content-center align-items-center flex-column">
-    <h1>BookStore</h1>
-    <a href="https://github.com/UzhKeksy/BookStore">Github</a>
-    <p class="text-muted">v.01</p>
-</div>
