--- app/views/DashboardView.php
@@ -1,3 +1,0 @@
-<div class="text-center">
-    Dashboard
-</div>
