+++ composer.json
@@ -9,7 +9,7 @@
   "require": {},
   "autoload": {
     "psr-4": {
-      "app\\": "app/"
+      "jarvis\\": "jarvis/"
     }
   }
 }
