+++ jarvis/core/Request.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace app\core;
+namespace jarvis\core;
 
 class Request
 {
