+++ jarvis/models/Model.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace app\models;
+namespace jarvis\models;
 
 abstract class Model implements IModel
 {
