+++ jarvis/models/ModelObject.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace app\models;
+namespace jarvis\models;
 
 class ModelObject
 {
