+++ jarvis/views/WelcomeView.php
@@ -1,0 +1,3 @@
+<div class="w-full h-screen flex justify-center items-center">
+    <h1 class="text7xl animate-pulse">Welcome</h1>
+</div>
